requirements/base.txt
# General
Django==1.9.12

# Map
django-geoposition==0.3.0

# Image
easy_thumbnails==2.3
Pillow==4.0.0

# Django Rest Framework
djangorestframework==3.6.2
markdown==2.6.8
django-filter==1.0.2
djoser==1.1.5
